Updated complex wrappers

+++ trunk/psi-jami/jami-xml/src/main/java/psidev/psi/mi/jami/xml/extension/AbstractXmlEntity.java
@@ -651,7 +651,7 @@ public abstract class AbstractXmlEntity<F extends Feature> implements ExtendedPs
                         // we recognize the stoichiometry range
                         if (stcs.length == 2){
                             try{
-                                XmlStoichiometry s = new XmlStoichiometry((long)Double.parseDouble(stc.trim()));
+                                XmlStoichiometry s = new XmlStoichiometry((int)Double.parseDouble(stc.trim()));
                                 s.setSourceLocator(sourceLocator);
                                 stoichiometry = s;
                                 return false;
@@ -672,7 +672,7 @@ public abstract class AbstractXmlEntity<F extends Feature> implements ExtendedPs
                     // simple stoichiometry
                     else {
                         try{
-                            XmlStoichiometry s = new XmlStoichiometry((long)Double.parseDouble(stc.trim()));
+                            XmlStoichiometry s = new XmlStoichiometry((int)Double.parseDouble(stc.trim()));
                             s.setSourceLocator(sourceLocator);
                             stoichiometry = s;
                             return false;
